requirements.txt
requests
pandas
scikit-learn
python-dotenv
Flask
gunicorn
numpy
joblib